Refactor package privacy of Email package part 2
